Drop idempotency table, ID prefix, and CLAUDE.md note
Completes the cleanup started in the prior commit. The idempotency lib
had no callers, so the schema, 'idm' ID prefix, and CLAUDE.md line that
described it as "applied at API boundaries" were all dead. Renamed the
remaining schema file to rate-limits.ts now that idempotency_keys is
gone, and added a migration to drop the table.
