+++ src/db/schema/index.ts
@@ -9,4 +9,4 @@ export * from './commitments';
 export * from './activity';
 export * from './magic-links';
 export * from './auth';
-export * from './idempotency';
+export * from './rate-limits';
